Метод getKittens() класса Lion возырвщен к изначальному виду Убраны цифры в названии классов Изменены названия методов и классов в едином стиле нейминга Объединена группировка тестов для классов Добавлена проверка эксепшн
