+++ src/test/java/com/example/DoesHaveManeParametrizedTest.java
@@ -7,7 +7,7 @@ import org.junit.runners.Parameterized;
 import static org.junit.Assert.assertEquals;
 
 @RunWith(Parameterized.class)
-public class DoesHaveManeTest {
+public class DoesHaveManeParametrizedTest {
 
     private String lionSex;
     private boolean expected;
