+++ src/test/java/com/example/FelineClassTest.java
@@ -5,10 +5,27 @@ import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.MockitoJUnitRunner;
+
+import java.util.List;
+import java.util.Arrays;
 import static org.junit.Assert.assertEquals;
 
 @RunWith(MockitoJUnitRunner.class)
-public class GetKittensTest {
+public class FelineClassTest {
+    @Test
+    public void shouldReturnTypeFood() throws Exception {
+        Feline feline = new Feline();
+        List<String> actual = feline.eatMeat();
+        List<String> expected = Arrays.asList("Животные", "Птицы", "Рыба");
+        assertEquals(expected, actual);
+    }
+    @Test
+    public void shouldReturnFamily() {
+        Feline feline = new Feline();
+        String actual = feline.getFamily();
+        String expected = "Кошачьи";
+        assertEquals(expected, actual);
+    }
     @Test
     public void shouldReturnQuantityKittens() {
         Feline feline = new Feline();
